requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
gdpc==6.0.2
idna==3.4
imageio==2.26.0
kiwisolver==1.4.4
matplotlib==3.7.0
more-itertools==9.0.0
NBT==1.5.1
nbtlib==2.0.4
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
PyGLM==2.7.0
pyglm-typing==0.2.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.1
six==1.16.0
termcolor==2.2.0
tifffile==2023.2.3
urllib3==1.26.14
